pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "numpy>=2.2.6",
    "openai>=1.82.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "plotly>=6.1.1",
    "psycopg2-binary>=2.9.10",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "streamlit>=1.45.1",
]
